+++ crates/client/db/src/tests/test_open.rs
@@ -1,6 +1,5 @@
 use super::common::*;
 use crate::DatabaseService;
-use mc_metrics::MetricsRegistry;
 use mp_chain_config::ChainConfig;
 
 #[tokio::test]
@@ -13,9 +12,8 @@ async fn test_open_different_chain_id() {
     let temp_dir = tempfile::TempDir::new().unwrap();
     {
         let chain_config = std::sync::Arc::new(ChainConfig::starknet_integration());
-        let _db =
-            DatabaseService::new(temp_dir.path(), None, false, chain_config, &MetricsRegistry::dummy()).await.unwrap();
+        let _db = DatabaseService::new(temp_dir.path(), None, false, chain_config).await.unwrap();
     }
     let chain_config = std::sync::Arc::new(ChainConfig::madara_test());
-    assert!(DatabaseService::new(temp_dir.path(), None, false, chain_config, &MetricsRegistry::dummy()).await.is_err());
+    assert!(DatabaseService::new(temp_dir.path(), None, false, chain_config).await.is_err());
 }