+++ crates/client/eth/src/state_update.rs
@@ -61,19 +61,19 @@ pub fn update_l1(
     //
     // TODO: Remove this check when the L1StateUpdate is properly verified
     if state_update.block_number > 500000u64 || chain_id.to_felt() == MAIN_CHAIN_ID {
-        log::info!(
+        tracing::info!(
             "🔄 Updated L1 head #{} ({}) with state root ({})",
             state_update.block_number,
             trim_hash(&state_update.block_hash),
             trim_hash(&state_update.global_root)
         );
 
-        block_metrics.l1_block_number.set(state_update.block_number as f64);
+        block_metrics.l1_block_number.record(state_update.block_number, &[]);
 
         backend
             .write_last_confirmed_block(state_update.block_number)
             .context("Setting l1 last confirmed block number")?;
-        log::debug!("update_l1: wrote last confirmed block number");
+        tracing::debug!("update_l1: wrote last confirmed block number");
     }
 
     Ok(())
@@ -86,9 +86,9 @@ pub async fn state_update_worker(
 ) -> anyhow::Result<()> {
     // Clear L1 confirmed block at startup
     backend.clear_last_confirmed_block().context("Clearing l1 last confirmed block number")?;
-    log::debug!("update_l1: cleared confirmed block number");
+    tracing::debug!("update_l1: cleared confirmed block number");
 
-    log::info!("🚀 Subscribed to L1 state verification");
+    tracing::info!("🚀 Subscribed to L1 state verification");
     // ideally here there would be one service which will update the l1 gas prices and another one for messages and one that's already present is state update
     // Get and store the latest verified state
     let initial_state = get_initial_state(eth_client).await.context("Getting initial ethereum state")?;
@@ -109,7 +109,6 @@ mod eth_client_event_subscription_test {
 
     use alloy::{node_bindings::Anvil, providers::ProviderBuilder, sol};
     use mc_db::DatabaseService;
-    use mc_metrics::{MetricsRegistry, MetricsService};
     use mp_chain_config::ChainConfig;
     use rstest::*;
     use tempfile::TempDir;
@@ -165,14 +164,13 @@ mod eth_client_event_subscription_test {
 
         // Initialize database service
         let db = Arc::new(
-            DatabaseService::new(&base_path, backup_dir, false, chain_info.clone(), &MetricsRegistry::dummy())
+            DatabaseService::new(&base_path, backup_dir, false, chain_info.clone())
                 .await
                 .expect("Failed to create database service"),
         );
 
         // Set up metrics service
-        let prometheus_service = MetricsService::new(true, false, 9615).unwrap();
-        let l1_block_metrics = L1BlockMetrics::register(prometheus_service.registry()).unwrap();
+        let l1_block_metrics = L1BlockMetrics::register().unwrap();
 
         let rpc_url: Url = anvil.endpoint().parse().expect("issue while parsing");
         let provider = ProviderBuilder::new().on_http(rpc_url);