+++ crates/node/src/service/rpc.rs
@@ -4,7 +4,6 @@ use jsonrpsee::server::ServerHandle;
 use tokio::task::JoinSet;
 
 use mc_db::DatabaseService;
-use mc_metrics::MetricsRegistry;
 use mc_rpc::{providers::AddTransactionProvider, versioned_rpc_api, Starknet};
 use mp_chain_config::ChainConfig;
 use mp_utils::service::Service;
@@ -27,7 +26,6 @@ impl RpcService {
         config: &RpcParams,
         db: &DatabaseService,
         chain_config: Arc<ChainConfig>,
-        metrics_handle: &MetricsRegistry,
         add_txs_method_provider: Arc<dyn AddTransactionProvider>,
     ) -> anyhow::Result<Self> {
         if config.rpc_disabled {
@@ -39,13 +37,14 @@ impl RpcService {
             (RpcMethods::Unsafe, _) => (true, true),
             (RpcMethods::Auto, false) => (true, true),
             (RpcMethods::Auto, true) => {
-                log::warn!(
+                tracing::warn!(
                     "Option `--rpc-external` will hide node operator endpoints. To enable them, please pass \
                      `--rpc-methods unsafe`."
                 );
                 (true, false)
             }
         };
+
         let (read, write, trace, internal, ws) = (rpcs, rpcs, rpcs, node_operator, rpcs);
         let starknet = Starknet::new(Arc::clone(db.backend()), chain_config.clone(), add_txs_method_provider);
         let metrics = RpcMetrics::register(metrics_handle)?;