+++ crates/tests/src/lib.rs
@@ -187,7 +187,6 @@ impl MadaraCmdBuilder {
                 self.args
                     .into_iter()
                     .chain([
-                        "--no-prometheus".into(),
                         "--base-path".into(),
                         format!("{}", self.tempdir.as_ref().display()),
                         "--rpc-port".into(),